==> 2_DataInfo.R <==
##########################################################################################################
### Script 2: Dataset Information ####
### This scripts takes three datasets (GUI, PSID, MCS)
### and provides information for paper and
### saves them for further analyses 
##########################################################################################################

##########################################################################################################
# 1. Load Libraries ######################################################################################
##########################################################################################################
library("tidyverse")
library("foreign")

##########################################################################################################
# 2. Lod Data ############################################################################################
##########################################################################################################
data_gui <- read.csv(file="../data/cleaned/1_1_GUI_dataset.csv")
data_psid <- read.csv(file="../data/cleaned/1_2_PSID_dataset.csv")
data_mcs <- read.csv(file="../data/cleaned/1_3_MCS_dataset.csv")

##########################################################################################################
# 3. GUI #################################################################################################
##########################################################################################################
nrow(data_gui)
table(data_gui$sex)
table(data_gui$age)

##########################################################################################################
# 4. PSID ################################################################################################
##########################################################################################################
nrow(data_psid)
table(data_psid$sex)
table(data_psid$age)

##########################################################################################################
# 5. MCS ## ##############################################################################################
##########################################################################################################
nrow(data_mcs)
table(data_mcs$sex)
table(data_mcs$age)

##########################################################################################################
# 6. GUI and PSID participants ###########################################################################
##########################################################################################################
nrow(data_gui) + nrow(data_psid) 

##########################################################################################################
# 7. Total participants ##################################################################################
##########################################################################################################
nrow(data_gui) + nrow(data_psid) + nrow(data_mcs)
