==> 3a_RunSCA.R <==
##########################################################################################################
### Script 3: Run SCAs and permutations ####
### This scripts takes three datasets (GUI, PSID, MCS)
### and runs SCAs using regression (not structural equation modelling)
### It consistes of multiple steps 
### 1. Loading libraries, controlling analyses and loading data
### 2. Running SCAs via three main functions get_data, get_model, get_coef
### 3. Make null models for bootstrapping
### 4. Run bootstrap models
### 5. Analyse bootstrap models 
### The output includes three specification curve analyses and one csv file of the bootstrap tests
##########################################################################################################

##########################################################################################################
# 1. Loading libraries, controlling analyses and loading data ############################################
##########################################################################################################

#####################################################################################
# a) Load libraries ####
#####################################################################################
library("tidyverse")
library("foreign")
library("lavaan")
library("heplots")
set.seed(2018)

#####################################################################################
# b) Turn on analyses ####
# 1 = yes, 0 = no
#####################################################################################
run_sca <- 1
run_null <- 1
run_boot <- 1

#####################################################################################
# c) Determine how many bootstraps to run
#####################################################################################
bootstraps <- 500

#####################################################################################
# d) Load data ####
#####################################################################################
data_gui <- read.csv(file="../data/cleaned/1_1_GUI_dataset.csv")
data_psid <- read.csv(file="../data/cleaned/1_2_PSID_dataset.csv")
data_mcs <- read.csv(file="../data/cleaned/1_3_MCS_dataset.csv")

##########################################################################################################
# 2. Running SCAs via three main functions, get_data, get_model and get_coef #############################
##########################################################################################################

#####################################################################################
#### Function: get_data
#### This function takes the specification and makes the apporiate dataset
####
#### Input: 
#### results_frame = a data frame showing specifications on each line
#### dataset = a character saying what dataset the specifications are for
#### data = the data of interest
####
#### Output: A list with the folliwng items
#### results_frame = retain results_frame to use in next function
#### dataset = retain dataset to use in next function
#### data = dataset of interest, now with new dv and iv from specifications
####        furthermore for the bootstraps sample the data is bootstrapped
####
#### Method: It goes to the row of interest (i) of the results frame.
#### There it reads what the relevant iv and dv variables are and extracts these
#### from the dataset of interest.
#####################################################################################
get_data <- function(results_frame, dataset, data, boot) {
  
  # Setup variable names for technology use
  if (results_frame$tech[i] == "sr_tech") {
    data$iv <- data$sr_tech
  } else if (results_frame$tech[i]  == "tech_weekday_d") {
    data$iv <- data$tech_di_wd
  } else if (results_frame$tech[i]  == "tech_weekend_d") {
    data$iv <- data$tech_di_we
  } else if (results_frame$tech[i]  == "tech_weekday_c") {
    data$iv <- data$tech_cont_wd
  } else if (results_frame$tech[i]  == "tech_weekend_c") {
    data$iv <- data$tech_cont_we
  } else if (results_frame$tech[i]  == "tech_2hr_wd") {
    data$iv <- data$tech_2hr_wd
  } else if (results_frame$tech[i]  == "tech_1hr_wd") {
    data$iv <- data$tech_1hr_wd 
  } else if (results_frame$tech[i]  == "tech_30m_wd") {
    data$iv <- data$tech_30m_wd
  } else if (results_frame$tech[i]  == "tech_2hr_we") {
    data$iv <- data$tech_2hr_we 
  } else if (results_frame$tech[i]  == "tech_1hr_we") {
    data$iv <- data$tech_1hr_we
  } else if (results_frame$tech[i]  == "tech_30m_we") {
    data$iv <- data$tech_30m_we
  }
  
  # MSetup variable names for wellbeing
  if (dataset == "psid"){
    if (results_frame$measure[i] == "selfesteem") {
      data$dv <- data$selfesteem
    } else if (results_frame$measure[i] == "wellbeing") {
      data$dv <- data$wellbeing
    } 
    
  } else if (dataset == "gui") { #gui
    if (results_frame$measure[i] == "wellbeing") {
      data$dv <- data$wellbeing_measure
    } else if (results_frame$measure[i] == "sdq") {
      data$dv <- data$sdqtot
    } 
    
  } else {
    if (results_frame$measure[i] == "selfesteem") {
      data$dv <- data$selfesteem
    } else if (results_frame$measure[i] == "wellbeing") {
      data$dv <- data$wellbeing
    } else if (results_frame$measure[i] == "sdq") {
      data$dv <- data$sdqtot
    }
  }
  
  # Do bootstrapping
  if(boot == TRUE){
    if (b <= bootstraps){
      k=sample(nrow(data),replace=T)
      data <- data[k, ]
    } else {}
  } else {
    
  }

  #make output into list
  get_data_list <- list(results_frame, dataset, data)
  return(get_data_list)
}


#####################################################################################
#### Function: get_model (2 versions)
#### There are two versions: get_model_scale and get_model_noscale
#### _scale standardises the variables before computing a regression
#### _noscale does not standardise the variables before computing the regression (for bootstrapping)
####
#### Input: A list from the previous function containing:
#### results_frame = a data frame showing specifications on each line
#### dataset = a character saying what dataset the specifications are for
#### data = the data of interest with relevant iv and dv
####
#### Output: A list with the folliwng items
#### reg = regression model from the specification
#### results_frame = retain results_frame to use in next function
#### data = dataset of interest
####
#### Method: It goes to the row of interest (i) of the results frame.
#### There it reads what the relevant dataset and control variables,
#### this determines the model which is then run on the dataset of interest
#####################################################################################
get_model_scale <- function(get_data_list){
  
  #unpackage list
  results_frame <- get_data_list[[1]]
  dataset <- get_data_list[[2]]
  data <- get_data_list[[3]]
  
  # Model with control
  if (results_frame$control[i] == 1) {
    if (dataset == "mcs") {
      reg <- lm(
        #iv, gender, age, educational motivation
        scale(dv) ~ scale(iv) + scale(sex) + scale(age) + scale(edumot) + 
          #ethnicity, closeness to parents, natural father live w/ you
          scale(fd06e00) + scale(clpar) + scale(fcpaab00) + 
          #parent word activity score, parent highest academic qualification, employment of parent
          scale(fpwrdscm) + scale(fdacaq00) + scale(fd05s00) + 
          #child long-term illness, amount time pri. caretaker w/ child, Kessler scale caretaker, num. sibs in houshold
          scale(fpclsi00) + scale(fpchti00) + scale(fdkessl) + scale(fdtots00) + 
          # equivalised weekly family income
          scale(foede000), 
        data = data
      )
      
    } else {
      reg <- lm(
        scale(dv) ~ scale(iv) + scale(sex) + scale(age),
        data = data
      )
    }
  } else {
    reg <- lm(
      scale(dv) ~ scale(iv),
      data = data
    )
  }
  
  # Return model and results_frame
  get_model_list <- list(reg, results_frame, data)
  return(get_model_list)
}

get_model_noscale <- function(get_data_list){
  
  #unpackage list
  results_frame <- get_data_list[[1]]
  dataset <- get_data_list[[2]]
  data <- get_data_list[[3]]
  
  # Model with control
  if (results_frame$control[i] == 1) {
    if (dataset == "mcs") {
      reg <- lm(
        #iv, gender, age, educational motivation
        dv ~ iv + sex + age + edumot + 
          #ethnicity, closeness to parents, natural father live w/ you
          fd06e00 + clpar + fcpaab00 + 
          #parent word activity score, parent highest academic qualification, employment of parent
          fpwrdscm + fdacaq00 + fd05s00 + 
          #child long-term illness, amount time pri. caretaker w/ child, Kessler scale caretaker, num. sibs in houshold
          fpclsi00 + fpchti00 + fdkessl + fdtots00 + 
          # equivalised weekly family income
          foede000, 
        data = data
      )
      
    } else {
      reg <- lm(
        dv ~ iv + sex + age,
        data = data
      )
    }
  } else {
    reg <- lm(
      dv ~ iv,
      data = data
    )
  }
  
  # Return model and results_frame
  get_model_list <- list(reg, results_frame, data)
  return(get_model_list)
}

#####################################################################################
#### Function: get_coef
#### This function extracts variables of interest from specification regression
####
#### Input: A list from the previous function with the following items
#### reg = regression model from the specification
#### results_frame = retain results_frame to use in next function
#### data = dataset of interest
####
#### Output:
#### results_frame[i,] = the relevant line of the results frame with key variables
####
#### Method: The function takes the relevant regression and results frame.
#### It extracts the t_value, regression coefficiant, p_value, SE and N
#### It adds these to the relevant results frame
#####################################################################################
get_coef <- function(get_model_list) {
  
  #unpackage list
  reg <- get_model_list[[1]]
  results_frame <- get_model_list[[2]]
  
  results_frame$t_value[i] <-
    summary(reg)$coef[[2, 3]] %>% {
      ifelse(. == 0, NA, .)
    }
  results_frame$effect[i] <-
    summary(reg)$coef[[2, 1]] %>% {
      ifelse(. == 0, NA, .)
    }
  results_frame$p_value[i] <- summary(reg)$coef[[2, 4]]
  results_frame$standard_error[i] <-
    summary(reg)$coef[[2, 2]] %>% {
      ifelse(. == 0, NA, .)
    }
  results_frame$number[i] <- nobs(reg)
  results_frame$rsqrd[i] <- etasq(reg)["scale(iv)", "Partial eta^2"]
  return(results_frame[i,])
}

#####################################################################################
# a) Setup Results Frame ####
# Furthermore we add an additional results frame (suffix = boot) 
# for our bootstrapping models later in the code
#####################################################################################

###############################
#### Function: get_results_frame
#### Makes a data frame to add
#### results of SCA in
####
#### Input: 
#### measure = measures of wellbeing
#### tech = measures of tech
#### controls = relevant controls
####
#### Output:
#### results_frame = data frame
####
#### Method: it makes grid with
#### every possible combo of 
#### input variables
###############################

get_results_frame <- function(measure, tech, control){
  results_frame <- expand.grid(measure, tech, control)
  names(results_frame) <- c("measure", "tech", "control")
  results_frame[, c("t_value", "effect", "p_value", "standard_error", "number", "rsqrd")] <- NA
  return(results_frame)
}

###############################
# i) Define measures
###############################
### definewellbeing measures for each dataset
measure_gui <- c("sdq", "wellbeing")
measure_psid <- c("selfesteem", "wellbeing")
measure_mcs <- c("selfesteem", "wellbeing", "sdq")

### defince tech and controls (same for all datasets)
tech <- c("sr_tech", 
          "tech_weekday_d", "tech_weekend_d", 
          "tech_weekday_c", "tech_weekend_c", 
          "tech_2hr_wd", "tech_1hr_wd", "tech_30m_wd", 
          "tech_2hr_we", "tech_1hr_we", "tech_30m_we")
control <- c(0, 1)

###############################
# ii) Run functions
###############################
results_frame_gui <- get_results_frame(measure_gui, tech, control)
results_frame_psid <- get_results_frame(measure_psid, tech, control)
results_frame_mcs <- get_results_frame(measure_mcs, tech, control)

###############################
# iii) Make bootstrap version
###############################
results_frame_gui_boot <- results_frame_gui
results_frame_psid_boot <- results_frame_psid 
results_frame_mcs_boot <- results_frame_mcs

#####################################################################################
# b) Run Specification Curve Analyses ####
#####################################################################################
sca_gui <- list(0)
sca_psid <- list(0)
sca_mcs <- list(0)

if (run_sca == 1){

  for (b in 1:(bootstraps+1)) {
    print(paste0("gui_bootstrap_",b))  
  for (i in 1:nrow(results_frame_gui)) {
    results_frame_gui[i,] <- get_coef(get_model_scale(get_data(
      results_frame = results_frame_gui, dataset = "gui", data = data_gui, boot = TRUE
    )))
    }
    sca_gui[[b]] <- results_frame_gui
    }

  for (b in 1:(bootstraps+1)) {
    print(paste0("psid_bootstrap_",b))    
  for (i in 1:nrow(results_frame_psid)) {
    results_frame_psid[i,] <- get_coef(get_model_scale(get_data(
      results_frame = results_frame_psid, dataset = "psid", data = data_psid, boot = TRUE
    )))
  }
    sca_psid[[b]] <- results_frame_psid
  }
  
  for (b in 1:(bootstraps+1)) {
      print(paste0("mcs_bootstrap_",b))  
  for (i in 1:nrow(results_frame_mcs)) {
    results_frame_mcs[i,] <- get_coef(get_model_scale(get_data(
      results_frame = results_frame_mcs, dataset = "mcs", data = data_mcs, boot = TRUE
    )))
  }
    sca_mcs[[b]] <- results_frame_mcs
  }
  
  # Save SCAs
  write.csv(sca_gui[[bootstraps+1]], file="../data/objects/3_1_GUI_SCA.csv")
  write.csv(sca_psid[[bootstraps+1]], file="../data/objects/3_2_PSID_SCA.csv")
  write.csv(sca_mcs[[bootstraps+1]], file="../data/objects/3_3_MCS_SCA.csv")
  
  write_rds(sca_gui, "./data/objects/3_1_GUI_SCA_SCAboot.rds")
  write_rds(sca_psid, "./data/objects/3_2_PSID_SCA_SCAboot.rds")
  write_rds(sca_mcs, "./data/objects/3_3_MCS_SCA_SCAboot.rds")
  
} else {
  
  # Read SCAs
  results_frame_gui <- read.csv(file="../data/objects/3_1_GUI_SCA.csv")
  results_frame_psid <- read.csv(file="../data/objects/3_2_PSID_SCA.csv")
  results_frame_mcs <- read.csv(file="../data/objects/3_3_MCS_SCA.csv")
  
  sca_gui <- read_rds("../data/objects/3_1_GUI_SCAboot.rds")
  sca_psid <- read_rds("../data/objects/3_2_PSID_SCAboot.rds")
  sca_mcs <- read_rds("../data/objects/3_3_MCS_SCAboot.rds")
  
}

##########################################################################################################
# 3. Make Null Models for Bootstrapping ##################################################################
##########################################################################################################

#####################################################################################
#### Function: get_ynull
#### This function contrains model under null for each specification
####
#### Input: A list from the previous function with the following items
#### reg = regression model from the specification
#### results_frame = retain results_frame to use in next function
#### data = dataset of interest
####
#### Output:
#### y.null.i = dv data for specification constrained under null
####
#### Method: The function takes the relevant regression and results frame.
#### It extracts the regression coefficiant, and multiplies it by the iv
#### It substracts this from the dv creating a dataset where null is true
#####################################################################################
get_ynull <- function(get_model_list) {
  
  #unpackage list
  reg <- get_model_list[[1]]
  data <- get_model_list[[3]]
  
  # extract coefficient
  b.i <-
    summary(reg)$coef[[2, 1]] %>% {
      ifelse(. == 0, NA, .)
    }
  
  #make null model
  y.null.i <- data$dv-(b.i*data$iv)
  
  return(y.null.i)
}


###############################
# i) Run Null Models
###############################
y_null_gui <- list(0)
y_null_mcs <- list(0)
y_null_psid <- list(0)

if (run_null == 1){
  
  for (i in 1:nrow(results_frame_gui)) {
    print(paste0("GUI_NULL_", i))
    y_null_gui[[i]] <- get_ynull(get_model_noscale(get_data(
      results_frame = results_frame_gui, dataset = "gui", data = data_gui, boot = FALSE
    )))
  }
  
  for (i in 1:nrow(results_frame_psid)) {
    print(paste0("PSID_NULL_", i))
    y_null_psid[[i]] <- get_ynull(get_model_noscale(get_data(
      results_frame = results_frame_psid, dataset = "psid", data = data_psid, boot = FALSE
    )))
  }
  
  for (i in 1:nrow(results_frame_mcs)) {
    print(paste0("MCS_NULL_", i))
    y_null_mcs[[i]] <- get_ynull(get_model_noscale(get_data(
      results_frame = results_frame_mcs, dataset = "mcs", data = data_mcs, boot = FALSE
    )))
  }
  
  # Save null models
  write_rds(y_null_gui, "../data/objects/3_1_GUI_ynull.rds")
  write_rds(y_null_psid, "../data/objects/3_2_PSID_ynull.rds")
  write_rds(y_null_mcs, "../data/objects/3_3_MCS_ynull.rds")
  
} else {
  
  # Read null models
  y_null_gui <- read_rds("../data/objects/3_1_GUI_ynull.rds")
  y_null_psid <- read_rds("../data/objects/3_2_PSID_ynull.rds")
  y_null_mcs <- read_rds("../data/objects/3_3_MCS_ynull.rds")
  
}


##########################################################################################################
# 4. Run Bootstrap Models ################################################################################
##########################################################################################################

#####################################################################################
#### Function: get_boot_data
#### This function takes the specification and makes the apporiate bootstrapped dataset
#### Except the last round trough (b >= bootstraps) it runs the original function
####
#### Input: 
#### results_frame = a data frame showing specifications on each line
#### dataset = a character saying what dataset the specifications are for
#### data = the data of interest
#### y.null = list of null datasets for each specification
####
#### Output: A list with the folliwng items
#### results_frame = retain results_frame to use in next function
#### dataset = retain dataset to use in next function
#### data = dataset of interest, now with new dv and iv for specification constrained under null
####
#### Method: It goes to the row of interest (i) of the results frame.
#### There it reads what the relevant iv and dv variables are and extracts these
#### from the dataset of interest.
#### During the bootstrapping tests (b <= bootstraps) it makes a bootstrapped dataset
#### that is constrained under the null. In the last test, it makes non-bootstrapped dataset
#####################################################################################
get_boot_data <- function(results_frame, dataset, data, y.null) {
  
  # Setup variable names for technology use
  
    if (results_frame$tech[i] == "sr_tech") {
      data$iv <- data$sr_tech
    } else if (results_frame$tech[i]  == "tech_weekday_d") {
      data$iv <- data$tech_di_wd
    } else if (results_frame$tech[i]  == "tech_weekend_d") {
      data$iv <- data$tech_di_we
    } else if (results_frame$tech[i]  == "tech_weekday_c") {
      data$iv <- data$tech_cont_wd
    } else if (results_frame$tech[i]  == "tech_weekend_c") {
      data$iv <- data$tech_cont_we
    } else if (results_frame$tech[i]  == "tech_2hr_wd") {
      data$iv <- data$tech_2hr_wd
    } else if (results_frame$tech[i]  == "tech_1hr_wd") {
      data$iv <- data$tech_1hr_wd 
    } else if (results_frame$tech[i]  == "tech_30m_wd") {
      data$iv <- data$tech_30m_wd
    } else if (results_frame$tech[i]  == "tech_2hr_we") {
      data$iv <- data$tech_2hr_we 
    } else if (results_frame$tech[i]  == "tech_1hr_we") {
      data$iv <- data$tech_1hr_we
    } else if (results_frame$tech[i]  == "tech_30m_we") {
      data$iv <- data$tech_30m_we
    }
  
  # Setup variable names for wellbeing
  if (b <= bootstraps){
    data$dv <- y.null[[i]]
  } else {
    if (dataset == "psid"){
    if (results_frame$measure[i] == "selfesteem") {
      data$dv <- data$selfesteem
    } else if (results_frame$measure[i] == "wellbeing") {
      data$dv <- data$wellbeing
    } 
    
  } else if (dataset == "gui") { #gui
    if (results_frame$measure[i] == "wellbeing") {
      data$dv <- data$wellbeing_measure
    } else if (results_frame$measure[i] == "sdq") {
      data$dv <- data$sdqtot
    } 
    
  } else {
    if (results_frame$measure[i] == "selfesteem") {
      data$dv <- data$selfesteem
    } else if (results_frame$measure[i] == "wellbeing") {
      data$dv <- data$wellbeing
    } else if (results_frame$measure[i] == "sdq") {
      data$dv <- data$sdqtot
    }
  }}
  
  # Do bootstrapping
  if (b <= bootstraps){
  k=sample(nrow(data),replace=T)
  data <- data[k, ]
  } else {}
  
  #make output into list
  get_data_list <- list(results_frame, dataset, data)
  return(get_data_list)
}

#####################################################################################
# a) Run Bootstrapping ####
#####################################################################################
bootstraps_gui <- list(0)
bootstraps_psid <- list(0)
bootstraps_mcs <- list(0)

if (run_boot == 1){
  
for (b in 1:(bootstraps+1)) {
  print(paste0("gui_bootstrap_",b))
  results_frame_gui_it <- results_frame_gui_boot
  for (i in 1:nrow(results_frame_gui_it)) {
    results_frame_gui_it[i,] <- get_coef(get_model_noscale(get_boot_data(
      results_frame = results_frame_gui_it, dataset = "gui", data = data_gui, y.null = y_null_gui
    )))
  }
  bootstraps_gui[[b]] <- results_frame_gui_it
}

for (b in 1:(bootstraps+1)) {
  print(paste0("psid_bootstrap_",b))
  results_frame_psid_it <- results_frame_psid_boot
  for (i in 1:nrow(results_frame_psid_it)) {
    results_frame_psid_it[i,] <- get_coef(get_model_noscale(get_boot_data(
      results_frame = results_frame_psid_it, dataset = "psid", data = data_psid, y.null = y_null_psid
    )))
  }
  bootstraps_psid[[b]] <- results_frame_psid_it
}

for (b in 1:(bootstraps+1)) {
  print(paste0("mcs_bootstrap_",b))
  results_frame_mcs_it <- results_frame_mcs_boot
  for (i in 1:nrow(results_frame_mcs_it)) {
    results_frame_mcs_it[i,] <- get_coef(get_model_noscale(get_boot_data(
      results_frame = results_frame_mcs_it, dataset = "mcs", data = data_mcs, y.null = y_null_mcs
    )))
  }
  bootstraps_mcs[[b]] <- results_frame_mcs_it
}
  
  write_rds(bootstraps_gui, "../data/objects/3_1_GUI_boot.rds")
  write_rds(bootstraps_psid, "../data/objects/3_2_PSID_boot.rds")
  write_rds(bootstraps_mcs, "../data/objects/3_3_MCS_boot.rds")
  
} else {
  
  # Read null models
  bootstraps_gui <- read_rds("../data/objects/3_1_GUI_boot.rds")
  bootstraps_psid <- read_rds("../data/objects/3_2_PSID_boot.rds")
  bootstraps_mcs <- read_rds("../data/objects/3_3_MCS_boot.rds")
  
}
